+++ src/event-listener.d.ts
@@ -1,2 +1,7 @@
-interface EventListener {
+interface EventListenerIntf {
+  handleEvent(event: any): void;
 }
+
+type EventListenerFunc = (event: any) => void;
+
+type EventListener = EventListenerIntf | EventListenerFunc;